MdfParser update 1
This submission updates MdfParser:
- all serialization methods now take a Version parameter as input
  => next update will make use of these
- all serialization methods are now static
  => no longer any need to create instances of IO objects when serializing
- code cleanup
- removed deprecated methods from SAX2Parser


git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@1832 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Server/src/Services/Feature/ServerFeatureService.cpp
@@ -1198,7 +1198,7 @@ MgByteReader* MgServerFeatureService::GetWfsFeature(MgResourceIdentifier* fs,
         //create a modifiable copy of the filter
         STRING ogcFilter = wfsFilter;
         STRING GEOM_PROP_TAG = L"%MG_GEOM_PROP%"; //NOXLATE
-        int geomPropPos = 0;
+        size_t geomPropPos = 0;
         while((geomPropPos = ogcFilter.find(GEOM_PROP_TAG, geomPropPos)) > 0)
         {
             ogcFilter = ogcFilter.replace(geomPropPos, GEOM_PROP_TAG.length(), fc->GetDefaultGeometryPropertyName());
